in order to use returningResult in JOOQ, I need customerID to be auto-increment; and in order to make it auto-increment it should not be Decimal in mySQL

+++ water-voucher/src/main/java/com/wv/controllers/CustomerControlller.java
@@ -34,7 +34,7 @@ public class CustomerControlller {
 	}
 	
 	@GetMapping("employees/{id}")
-	public Customer get(@PathVariable Long id) {
+	public Customer get(@PathVariable Integer id) {
 		return customerRepoJooqImpl.get(id).orElse(null);
 	}
 }
